requirements.txt
flask==3.0.0
openai==1.52.0
python-dotenv==1.0.0
pydantic==2.10
requests==2.31.0
gunicorn==21.2.0
flask-limiter==3.5.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
python-decouple==3.8 
crawl4ai==0.6.3